Make pyOpenSSL an optional dependency
See Sage trac #13385: http://trac.sagemath.org/sage_trac/ticket/13385

Basically we want to fail gracefully if pyOpenSSL isn't installed. This
commit preemptively turns off secure mode as soon as notebook() is
called if it detects pyOpenSSL is not installed. There is also a single
OpenSSL import in sagenb's own code, so I wrapped that as well just in
case.