Get rid of custom SageNBFlask.save_session
This hack probably hasn't been needed for a while but now it's started
actively breaking so it's time to get rid of it. The docstring of the
function says "This method needs to stay in sync with the version in
Flask", but it hasn't been since 2011 [1].

This change shouldn't affect behavior - the value of the `httponly` flag
that's eventually passed to a newer incarnation of `save_cookie` is read
from the app object's config pseudodictionary. See [2] and [3].

[1] https://github.com/mitsuhiko/flask/blame/0.10.1/flask/app.py#L837
[2] https://github.com/mitsuhiko/flask/blob/0.10.1/flask/sessions.py#L323
[3] https://github.com/mitsuhiko/flask/blob/0.10.1/flask/sessions.py#L233