Update notebook.py
This fix is to address the issue [discussed on the mailing list](https://groups.google.com/d/topic/sage-support/EK9nOmx4lsQ/discussion).
In Sage-7.4 the Notebook cannot read the text files with UTF-8 encoding (and perhaps, other 8-bit encodings too).

The idea of the proposed patch is first to try to decode the text as UTF-8, and if it does not work, fall back to Latin-1 encoding. This is not a graceful solution indeed, and is definitely not a smart thing to do.
However, in this case even if the text file is not in Latin-1 encoding, but in some other (let's say, in cp1251), the Notebook will still print the file to the screen instead of just saying "500: Internal server error." In such case the user will _see_ that the problem is the encoding of the text file.