Updated built-in sphinxify routine to support proposed changes in Sage
This is a change needed in order to support changes I'm proposing to Sage in https://trac.sagemath.org/ticket/21732

Those changes haven't been approved yet though, so obviously this PR should not be merged until something is finalized in that Trac ticket.  Ideally, sagenb should probably not rely on Sage at all for this, and I'd be in favor of a more extensive PR for sagenb to just include its own Sphinx configuration files for this purpose.  But for now this is the simplest solution.