Add initialization to docstring doctest
The fix of an issue related to introspection in Sage, revealed an issue
with one of the doctests in sagenb. More specifically, before this
commit, `docstring` was called without first initializing the module,
but initialization is necessary so that `EMBEDDED_MODE` gets set, which
is needed to format docstrings correctly.

Apparently, this is just an issue with this specific doctest, and not
the notebook in general.

For more information, see https://trac.sagemath.org/ticket/26906.